add basic association support for ABAP CDS views
Associations for ABAP CDS are not materialized in the db, but old versions of AS ABAP had an ADT service to get the corresponding JOIN clause (to be used from the SQL console).
This commit maps a single association path into its corresponding JOIN behind the scenes to be able to use associations to a limited degree.
Path expression attributes, CTEs, and set operators are currently not supported but a query like the following should work without any problems:
select _spfli.airpfrom 
  from "SAPA4H"."DEMO_CDS_ASSOC_SCARR"